# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tıraş - Barber Appointment SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user (barber or customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with an email address or Turkish phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset by email or phone
      description: >-
        Returns the same response shape and message whether or not the account
        exists. The opaque resetId must also be indistinguishable for real and
        decoy requests so this endpoint cannot be used for account enumeration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Generic password-reset instructions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetRequestResponse"
        "400":
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many reset requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password-reset/verify:
    post:
      operationId: verifyPasswordReset
      tags: [auth]
      summary: Verify a password-reset code
      description: >-
        Verifies a six-digit, single-use code and returns a short-lived opaque
        reset token. Invalid, expired, and unknown reset attempts use the same
        error response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetVerifyRequest"
      responses:
        "200":
          description: Code verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetVerifyResponse"
        "400":
          description: Invalid or expired reset attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many verification attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password-reset/reset:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password with a reset token
      description: >-
        Consumes the short-lived reset token. A successfully consumed token
        cannot be reused.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmRequest"
      responses:
        "200":
          description: Password reset completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetMessageResponse"
        "400":
          description: Invalid or expired reset token, or invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many reset attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Barbers
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    RegisterRequest:
      oneOf:
        - $ref: "#/components/schemas/CustomerRegisterRequest"
        - $ref: "#/components/schemas/BarberRegisterRequest"
      discriminator:
        propertyName: role
        mapping:
          customer: "#/components/schemas/CustomerRegisterRequest"
          barber: "#/components/schemas/BarberRegisterRequest"
    CustomerRegisterRequest:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Email address; normalized to lowercase by the server.
          example: ayse@example.com
        password:
          type: string
          minLength: 6
          maxLength: 72
          writeOnly: true
        phone:
          $ref: "#/components/schemas/TurkishPhoneNumber"
        role:
          type: string
          enum: [customer]
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        name:
          type: string
          deprecated: true
          description: Legacy full-name alias; use firstName and lastName.
        surname:
          type: string
          deprecated: true
          description: Legacy surname alias; use lastName.
      required:
        - email
        - password
        - phone
        - role
        - firstName
        - lastName
    BarberRegisterRequest:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Email address; normalized to lowercase by the server.
          example: salon@example.com
        password:
          type: string
          minLength: 6
          maxLength: 72
          writeOnly: true
        phone:
          $ref: "#/components/schemas/TurkishPhoneNumber"
        role:
          type: string
          enum: [barber]
        businessName:
          type: string
          minLength: 1
          maxLength: 160
        authorizedFirstName:
          type: string
          minLength: 1
          maxLength: 100
        authorizedLastName:
          type: string
          minLength: 1
          maxLength: 100
        address:
          type: string
          minLength: 1
          maxLength: 500
        shopName:
          type: string
          deprecated: true
          description: Legacy business-name alias; use businessName.
        shopAddress:
          type: string
          deprecated: true
          description: Legacy address alias; use address.
        name:
          type: string
          deprecated: true
          description: Legacy authorized-person name alias.
        surname:
          type: string
          deprecated: true
          description: Legacy authorized-person surname alias.
      required:
        - email
        - password
        - phone
        - role
        - businessName
        - authorizedFirstName
        - authorizedLastName
        - address
    LoginRequest:
      oneOf:
        - $ref: "#/components/schemas/IdentifierLoginRequest"
        - $ref: "#/components/schemas/LegacyEmailLoginRequest"
...
    LegacyEmailLoginRequest:
      type: object
      deprecated: true
      additionalProperties: false
      description: Legacy email-based login body; use identifier instead.
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          deprecated: true
          description: Legacy alias for identifier.
        password:
          type: string
          minLength: 6
          maxLength: 72
          writeOnly: true
      required:
        - email
        - password
...
    PasswordResetRequest:
      type: object
      additionalProperties: false
      properties:
        identifier:
          type: string
          minLength: 1
          maxLength: 254
          description: >-
            Email address or a canonical Turkish mobile number: exactly 10
            digits starting with 5, without a leading zero, country code, or
            formatting.
          example: "5551112233"
        channel:
          $ref: "#/components/schemas/PasswordResetChannel"
          description: >-
            Preferred delivery channel. If omitted, the server infers it from
            the identifier.
      required:
        - identifier
    PasswordResetRequestResponse:
      type: object
      properties:
        message:
          type: string
          description: Generic instructions that do not reveal account existence.
        resetId:
          type: string
          minLength: 1
          description: >-
            Opaque identifier for the verification attempt. Real and decoy IDs
            are indistinguishable.
      required:
        - message
        - resetId
    PasswordResetVerifyRequest:
      type: object
      additionalProperties: false
      properties:
        resetId:
          type: string
          minLength: 1
          description: Opaque identifier returned by the reset request.
        code:
          type: string
          pattern: '^\d{6}$'
          minLength: 6
          maxLength: 6
          description: Six-digit, single-use verification code.
          example: "123456"
      required:
        - resetId
        - code
    PasswordResetVerifyResponse:
      type: object
      properties:
        message:
          type: string
        resetToken:
          type: string
          minLength: 1
          description: Opaque, short-lived, single-use password-reset token.
      required:
        - message
        - resetToken
    PasswordResetConfirmRequest:
      type: object
      additionalProperties: false
      properties:
        resetToken:
          type: string
          minLength: 1
          description: Opaque token returned after successful code verification.
        newPassword:
          type: string
          minLength: 6
          maxLength: 72
          writeOnly: true
      required:
        - resetToken
        - newPassword
    PasswordResetMessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [barber, customer]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
...